Reorganize tests Rename test_memory_blob.rb to test_blob.rb for consistence with last commit
